Created CONTRIBUTERS.md and added my name to it
As an exercise I would recommend you both create a CONTRIBUTERS.md for each of the remaining
two projects as well as all of us add our names to the CONTRIBUTERS.md files for each project.
This should familiarize you with the basic workflow of git and how merges work.

Best,
Nick